extract date calendar primitives to shared/calendar
