+++ packages/react-components/src/shared/calendar/index.ts
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2025 Palantir Technologies, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export {
+  buildDisabledMatchers,
+  DEFAULT_FROM_YEAR,
+  DEFAULT_TO_YEAR,
+  stopPropagation,
+} from "./calendarShared.js";
+export {
+  CALENDAR_COMPONENTS,
+  CLASS_NAMES,
+  type DateCalendarProps,
+  default as DateCalendar,
+} from "./DateCalendar.js";
+export {
+  type DateRangeCalendarProps,
+  default as DateRangeCalendar,
+} from "./DateRangeCalendar.js";
+export { LazyDateCalendar } from "./LazyDateCalendar.js";
+export { LazyDateRangeCalendar } from "./LazyDateRangeCalendar.js";
